Added more functionality to FoodClassifier and TestClassifier
FoodClassifier now has a confidence that can be used when returning the label.
set and get functions for FoodClassifier
Labels can now be a file, instead of a list.
FoodClassifier now makes its own resnet model. The extension model is held seperate.
Train_classifier_from_images now saves the extModel with a timestamp, so this should be unique and not overwrite other models.
Added an add_to_train function to add a new image to the training set.  this will append the feature values to the npy file of the given label. It will create a new npy file if none exists for the given label.